Split plugin and widget
